windows/network.c: refactor switch in sk_newlistener.
The code that diverges based on the address family is now in the form
of a switch statement, rather than an unwieldy series of chained ifs.
And the final call to bind() has all its arguments worked out in the
previous switch, rather than computing them at the last minute with an
equally unwieldy set of ?: operators that repeat the previous test.

This will make it easier to add more cases, and also, to keep each
case under its own ifdef without losing too much legibility.